Fix time tracker schema to support decimal effort points
Changed effort_points column from INTEGER to DECIMAL(5, 2) to support
fractional hours. This fixes the error "invalid input syntax for type
integer: '1.25'" when logging time entries with decimal hours.

Changes:
- Updated supabase-migration.sql to use DECIMAL for effort_points
- Added migration-fix-effort-points.sql for updating existing databases

User action required:
Run migration-fix-effort-points.sql in Neon SQL Editor to update the
database schema for existing installations.

+++ productivity-hub/supabase-migration.sql
@@ -40,7 +40,7 @@ CREATE TABLE time_entries (
   date DATE NOT NULL,
   category TEXT CHECK (category IN ('UK Job', 'Solar App', 'Factory App', 'Personal', 'Uni', 'Gym')) NOT NULL,
   hours DECIMAL(5, 2) NOT NULL,
-  effort_points INTEGER NOT NULL,
+  effort_points DECIMAL(5, 2) NOT NULL,
   description TEXT,
   created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW()
 );
